refactor(fabric): eliminate topology_solver_sat_session.hpp — no extra files
Fold TopologySatSession struct definition and all private function
declarations into the existing topology_solver_sat_solver.hpp.
TopologySatSession is defined locally in topology_solver_sat.cpp
(right after its only include of the private header), so no
additional header file is needed.

Delete topology_solver_sat_session.hpp and update the single
consumer (test_topology_sat_encoder.cpp) to include the solver
header directly.
