+++ tt_metal/fabric/topology_solver_sat_solver.hpp
@@ -5,6 +5,9 @@
 #pragma once
 
 #include <memory>
+#include <vector>
+
+#include <tt-metalium/experimental/fabric/topology_solver.hpp>
 
 namespace tt::tt_fabric::detail {
 
@@ -42,5 +45,36 @@ private:
     std::unique_ptr<Impl> impl_;
     int next_var_ = 0;
 };
+
+// Internal SAT function declarations — implemented in topology_solver_sat.cpp.
+
+bool topology_sat_encode_hard_constraints(
+    TopologySatSolver& solver,
+    const TopologySatGraphView& graph_data,
+    const TopologySatConstraintView& constraint_data,
+    TopologySatHardEncoding& enc,
+    ConnectionValidationMode validation_mode = ConnectionValidationMode::RELAXED);
+
+bool topology_sat_decode_hard_solution(
+    TopologySatSolver& solver, const TopologySatHardEncoding& enc, std::vector<int>& mapping_out);
+
+bool topology_sat_add_blocking_clause_for_mapping(
+    TopologySatSolver& solver, TopologySatHardEncoding& enc, const std::vector<int>& raw_mapping, bool unique_shapes);
+
+// Template overload: converts GraphIndexData/ConstraintIndexData to views and delegates.
+template <typename TargetNode, typename GlobalNode>
+bool topology_sat_encode_hard_constraints(
+    TopologySatSolver& solver,
+    const GraphIndexData<TargetNode, GlobalNode>& graph_data,
+    const ConstraintIndexData<TargetNode, GlobalNode>& constraint_data,
+    TopologySatHardEncoding& enc,
+    ConnectionValidationMode validation_mode = ConnectionValidationMode::RELAXED) {
+    return topology_sat_encode_hard_constraints(
+        solver,
+        TopologySatGraphView(graph_data),
+        TopologySatConstraintView(constraint_data),
+        enc,
+        validation_mode);
+}
 
 }  // namespace tt::tt_fabric::detail
